Created a World Gen with Multiple Colors! :lock:
To manage that, I had to rewrite the way that chunks are created, it's using 3 geometries and 3 meshes for each chunk.
#WorldWorker3.js
The most basic world gen I could think of was the one with dirt and grass, easy to create, here's a preview:
#preview_v025.png "A world gen with a top layer of green grass"
Sadly, There will be a perfomance impact because of an extra mesh for each color for each chunk. It's definitely worth it, though!

#Chunk.js
It was very easy implementing the 3 different colors, all I had to do is create an array for every threejs object and put the 3 meshes on the scene, that simplicity was all thanks to the way the Chunk class was abstracted.

I'm starting to get really good at creating code that is easy to maintain.